requirements.txt
amqp==2.2.2
apturl==0.5.2
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
bcrypt==3.1.4
beautifulsoup4==4.4.1
billiard==3.5.0.3
blinker==1.3
Brlapi==0.6.4
celery==4.1.0
certifi==2019.3.9
cffi==1.11.5
chardet==2.3.0
checkbox-support==0.22
click==6.7
client==0.0.1
command-not-found==0.3
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
defer==1.0.6
devscripts===2.16.2ubuntu3
Django==2.0.4
feedparser==5.1.3
Flask==0.12.2
future==0.17.1
google==2.0.1
guacamole==0.9.2
gunicorn==19.9.0
html5lib==0.999
httplib2==0.9.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itsdangerous==0.24
jedi==0.13.3
Jinja2==2.8
kazam==1.4.5
kombu==4.1.0
language-selector==0.1
libaio==0.2
louis==2.6.4
lxml==3.5.0
Magic-file-extensions==0.2
Mako==1.0.3
MarkupSafe==0.23
mysql==0.0.2
mysql-connector==2.1.6
mysql-connector-python==2.0.4
mysqlclient==1.4.2.post1
ndg-httpsclient==0.5.1
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
paramiko==2.4.1
parsel==1.5.0
parso==0.3.4
pbr==5.1.3
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
pluggy==0.9.0
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.2.2
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
PyDispatcher==2.0.5
pygobject==3.20.0
PyHamcrest==1.9.0
PyJWT==1.3.0
PyMySQL==0.9.3
PyNaCl==1.2.1
pyOpenSSL==18.0.0
pyparsing==2.0.3
pysftp==0.2.9
python-apt==1.1.0b1+ubuntu0.16.4.5
python-dateutil==2.8.0
python-debian==0.1.27
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
python-systemd==231
pytz==2018.3
pyxdg==0.25
queuelib==1.5.0
RapidConnect==0.0.6
reportlab==3.3.0
requests==2.9.1
Scrapy==1.5.1
service-identity==17.0.0
sessioninstaller==0.0.0
six==1.10.0
stevedore==1.30.1
system-service==0.3
Twisted==18.7.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
uWSGI==2.0.18
vine==1.1.4
virtualenv==16.5.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
w3lib==1.19.0
websocket-client==0.51.0
Werkzeug==0.14.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zope.interface==4.5.0
